+++ tests/testthat/test-ao.R
@@ -1,5 +1,11 @@
 test_that("ao works", {
   himmelblau <- function(x) (x[1]^2 + x[2] - 11)^2 + (x[1] + x[2]^2 - 7)^2
+  ao_out <- ao(f = himmelblau, initial = c(0, 0))
+  checkmate::expect_list(ao_out, len = 5)
+})
+
+test_that("ao with custom gradient works", {
+  himmelblau <- function(x) (x[1]^2 + x[2] - 11)^2 + (x[1] + x[2]^2 - 7)^2
   gradient <- function(x) {
     c(
       4 * x[1] * (x[1]^2 + x[2] - 11) + 2 * (x[1] + x[2]^2 - 7),
@@ -10,6 +16,26 @@ test_that("ao works", {
   checkmate::expect_list(ao_out, len = 5)
   ao_random_out <- ao(f = himmelblau, initial = c(0, 0), partition = "random", gradient = gradient)
   checkmate::expect_list(ao_random_out, len = 5)
+})
+
+test_that("ao with custom Hessian works", {
+  f <- function(x) {
+    # print("huhu, my name is f")
+    x^2
+  }
+  g <- function(x) {
+    # print("huhu, my name is g")
+    2*x
+  }
+  h <- function(x) {
+    # print("huhu, my name is h")
+    matrix(2) # important to have matrix(2) instead of 2!
+  }
+  ao_out <- ao(
+    f = f, initial = 10, gradient = g, hessian = h,
+    base_optimizer = Optimizer$new("stats::nlm") # switch to nlm because stats::optim does not support Hessian
+  )
+  checkmate::expect_list(ao_out, len = 5)
 })
 
 test_that("ao with additional parameters works", {
@@ -46,6 +72,17 @@ test_that("ao with custom partition works", {
   checkmate::expect_list(ao_custom, len = 5)
 })
 
+test_that("ao with parameter bounds works", {
+  rosenbrock <- function(x) (1 - x[1])^2 + (x[2] - x[1]^2)^2
+  lower <- 1.5
+  upper <- 2.5
+  initial <- runif(2, min = lower, max = upper)
+  ao_bounds <- ao(f = rosenbrock, initial = initial, lower = lower, upper = upper)
+  checkmate::expect_list(ao_bounds, len = 5)
+  expect_true(all(ao_bounds$estimate <= upper))
+  expect_true(all(ao_bounds$estimate >= lower))
+})
+
 test_that("multiple ao processes work", {
   himmelblau <- function(x) (x[1]^2 + x[2] - 11)^2 + (x[1] + x[2]^2 - 7)^2
   out_multi <- ao(
@@ -59,4 +96,9 @@ test_that("multiple ao processes work", {
     add_details = FALSE
   )
   checkmate::expect_list(out_multi, len = 4)
+})
+
+test_that("ao fails graciously", {
+  f <- function(x) x[1]^2 + x[2]^2
+  expect_silent(ao(f = f, initial = 1)) # initial is mis-specified
 })